analytics-assistance/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-dotenv
langchain
langchain-openai
pydantic
